# backend/app/api/v1/ai_core.py
from typing import List, Optional
from fastapi import APIRouter, Depends, Query, HTTPException
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select

from app.core.database import get_db
from app.core.security import get_current_user
from app.models.models import User, CorridorEvent, Schedule, TMSDefect, SMMSFault, TDMSFault, BDMSRequest, TrackSection
from app.schemas.schemas import (
    SynthesizeRequest, CorridorEventSchema, CorridorEventDetailSchema,
    CriticalityScoreRequest, CriticalityExplainabilityResponse, ScheduleOptimizeRequest,
    ScheduleSchema, WhatIfSimulationRequest, WhatIfSimulationResponse, StandardResponse,
    ShadowBlockResultSchema, OpportunityWindowSchema, OpportunityMiningResponse,
    MaintenanceOpportunityTask
)
from app.services.spatial_synthesizer import synthesize_corridor_events, mine_maintenance_opportunities
from app.services.scoring import compute_criticality_scores, explain_event_criticality
from app.services.solver import optimize_maintenance_schedule
from app.services.shadow_blocking import execute_shadow_blocking, get_shadow_block_results
from app.services.simulator import run_what_if_simulation
from app.services.ws_manager import ws_manager

router = APIRouter(tags=["Stage 2 & 3 AI Decision Core & Scheduling"])

@router.post("/corridor-events/synthesize", response_model=StandardResponse[List[CorridorEventSchema]])
async def synthesize_events_endpoint(
    body: SynthesizeRequest,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    events = await synthesize_corridor_events(db, section_ids=body.section_ids)
    data = [CorridorEventSchema.model_validate(e) for e in events]
    return StandardResponse(data=data, meta={"total": len(data)})

@router.get("/corridor-events", response_model=StandardResponse[List[CorridorEventSchema]])
async def list_corridor_events(
    section_id: Optional[str] = Query(None),
    status: Optional[str] = Query(None),
    min_score: Optional[float] = Query(None),
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    stmt = select(CorridorEvent)
    if section_id:
        stmt = stmt.filter_by(section_id=section_id)
    if status:
        stmt = stmt.filter_by(status=status)
    if min_score is not None:
        stmt = stmt.filter(CorridorEvent.criticality_score >= min_score)
    res = await db.execute(stmt)
    events = res.scalars().all()
    data = [CorridorEventSchema.model_validate(e) for e in events]
    return StandardResponse(data=data, meta={"total": len(data)})

@router.get("/corridor-events/{id}", response_model=StandardResponse[CorridorEventDetailSchema])
async def get_corridor_event_detail(
    id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    res = await db.execute(select(CorridorEvent).filter_by(id=id))
    ev = res.scalar_one_or_none()
    if not ev:
        raise HTTPException(status_code=404, detail=f"CorridorEvent {id} not found")

    # Fetch source record details
    source_records = []
    if ev.merged_from:
        tms_res = await db.execute(select(TMSDefect).filter(TMSDefect.defect_id.in_(ev.merged_from)))
        for item in tms_res.scalars().all():
            source_records.append({"id": item.defect_id, "type": "TMS", "severity": item.severity, "desc": f"Track defect ({item.defect_type})"})
        
        smms_res = await db.execute(select(SMMSFault).filter(SMMSFault.fault_id.in_(ev.merged_from)))
        for item in smms_res.scalars().all():
            source_records.append({"id": item.fault_id, "type": "SMMS", "severity": item.severity, "desc": f"Signal fault ({item.fault_type})"})

        tdms_res = await db.execute(select(TDMSFault).filter(TDMSFault.fault_id.in_(ev.merged_from)))
        for item in tdms_res.scalars().all():
            source_records.append({"id": item.fault_id, "type": "TDMS", "severity": item.severity, "desc": f"Traction fault ({item.fault_type})"})

    detail = CorridorEventDetailSchema(
        id=ev.id,
        section_id=ev.section_id,
        departments=ev.departments,
        criticality_score=ev.criticality_score,
        confidence=ev.confidence,
        merged_from=ev.merged_from,
        status=ev.status,
        source_records=source_records
    )
    return StandardResponse(data=detail)

@router.post("/scoring/criticality", response_model=StandardResponse[List[CorridorEventSchema]])
async def score_criticality_endpoint(
    body: CriticalityScoreRequest,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    events = await compute_criticality_scores(db, event_ids=body.event_ids)
    data = [CorridorEventSchema.model_validate(e) for e in events]
    return StandardResponse(data=data, meta={"total": len(data)})

@router.get("/scoring/criticality/{event_id}", response_model=StandardResponse[CriticalityExplainabilityResponse])
async def get_criticality_explainability(
    event_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    score, breakdown = await explain_event_criticality(db, event_id)
    resp = CriticalityExplainabilityResponse(
        event_id=breakdown["event_id"],
        section_id=breakdown["section_id"],
        criticality_score=breakdown["criticality_score"],
        feature_breakdown=breakdown["feature_breakdown"],
        explanation=breakdown["explanation"]
    )
    return StandardResponse(data=resp)

@router.post("/schedule/optimize", response_model=StandardResponse[List[ScheduleSchema]])
async def optimize_schedule_endpoint(
    body: ScheduleOptimizeRequest,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    schedules = await optimize_maintenance_schedule(
        db,
        horizon_type=body.horizon_type,
        date_range=body.date_range,
        section_ids=body.section_ids
    )
    data = [ScheduleSchema.model_validate(s) for s in schedules]
    
    # Broadcast websocket update
    await ws_manager.broadcast({
        "event": "schedule_optimized",
        "count": len(data),
        "horizon_type": body.horizon_type
    })
    
    return StandardResponse(data=data, meta={"total": len(data)})

@router.post("/schedule/{schedule_batch_id}/shadow-block", response_model=StandardResponse[List[ScheduleSchema]])
async def shadow_block_endpoint(
    schedule_batch_id: str,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    updated = await execute_shadow_blocking(db, schedule_batch_id=schedule_batch_id)
    data = [ScheduleSchema.model_validate(s) for s in updated]
    return StandardResponse(data=data, meta={"message": f"Shadow-blocked {len(data)} maintenance schedules."})

@router.post("/simulate/what-if", response_model=StandardResponse[WhatIfSimulationResponse])
async def what_if_simulation_endpoint(
    body: WhatIfSimulationRequest,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    result = await run_what_if_simulation(
        db,
        schedule_batch_id=body.schedule_batch_id,
        variation_pct=body.variation_pct,
        iterations=body.iterations
    )
    resp = WhatIfSimulationResponse(**result)
    return StandardResponse(data=resp)

@router.get("/schedule/monthly", response_model=StandardResponse[List[ScheduleSchema]])
async def get_monthly_schedule(
    division: Optional[str] = Query(None),
    month: Optional[str] = Query(None),
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    stmt = select(Schedule).filter_by(horizon_type="monthly")
    res = await db.execute(stmt)
    schedules = res.scalars().all()
    data = [ScheduleSchema.model_validate(s) for s in schedules]
    return StandardResponse(data=data, meta={"total": len(data)})

@router.get("/schedule/weekly", response_model=StandardResponse[List[ScheduleSchema]])
async def get_weekly_schedule(
    section_id: Optional[str] = Query(None),
    week_start: Optional[str] = Query(None),
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    stmt = select(Schedule).filter_by(horizon_type="weekly")
    if section_id:
        stmt = stmt.filter_by(section_id=section_id)
    res = await db.execute(stmt)
    schedules = res.scalars().all()
    data = [ScheduleSchema.model_validate(s) for s in schedules]
    return StandardResponse(data=data, meta={"total": len(data)})

@router.post("/schedule/weekly/reoptimize", response_model=StandardResponse[List[ScheduleSchema]])
async def reoptimize_weekly_schedule(
    trigger_event_id: Optional[int] = Query(None),
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    schedules = await optimize_maintenance_schedule(db, horizon_type="weekly")
    data = [ScheduleSchema.model_validate(s) for s in schedules]
    return StandardResponse(data=data, meta={"message": "Weekly schedule re-optimized."})

@router.get("/schedule/{schedule_batch_id}", response_model=StandardResponse[List[ScheduleSchema]])
async def get_schedule_batch_detail(
    schedule_batch_id: str,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    res = await db.execute(select(Schedule))
    schedules = res.scalars().all()
    data = [ScheduleSchema.model_validate(s) for s in schedules]
    return StandardResponse(data=data, meta={"batch_id": schedule_batch_id, "total": len(data)})


@router.get("/shadow-blocks", response_model=StandardResponse[List[ShadowBlockResultSchema]])
async def get_shadow_block_details(
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    """
    USP 1 — Returns rich shadow block result objects with consolidation scores,
    downtime reduction estimates, task lists, and reasoning explanations.
    """
    results = await get_shadow_block_results(db)
    return StandardResponse(data=results, meta={"total": len(results)})


@router.get("/corridor-events/opportunities", response_model=StandardResponse[List[OpportunityMiningResponse]])
async def get_maintenance_opportunities(
    section_id: Optional[str] = Query(None),
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user)
):
    """
    USP 5 — Maintenance Opportunity Mining:
    Analyzes train schedule gaps to discover hidden maintenance windows.
    Returns scored opportunity windows with suitable task recommendations.
    """
    section_ids = [section_id] if section_id else None
    raw_opportunities = await mine_maintenance_opportunities(db, section_ids=section_ids)

    # Group by section_id
    grouped: dict = {}
    for opp in raw_opportunities:
        sid = opp["section_id"]
        if sid not in grouped:
            grouped[sid] = []
        grouped[sid].append(opp)

    responses = []
    for sid, opps in grouped.items():
        windows = []
        for o in opps:
            tasks = [
                MaintenanceOpportunityTask(
                    task_id=t["task_id"],
                    department=t["department"],
                    task_type=t["task_type"],
                    severity=t["severity"],
                    priority_score=t["priority_score"]
                )
                for t in o["suitable_tasks"]
            ]
            windows.append(OpportunityWindowSchema(
                window_id=o["window_id"],
                section_id=o["section_id"],
                corridor=o["corridor"],
                window_start=o["window_start"],
                window_end=o["window_end"],
                duration_minutes=o["duration_minutes"],
                traffic_level=o["traffic_level"],
                trains_in_window=o["trains_in_window"],
                suitable_tasks=tasks,
                block_suitability_score=o["block_suitability_score"],
                recommendation=o["recommendation"]
            ))

        high_score_wins = [w for w in windows if w.block_suitability_score >= 60]
        notes = (
            f"Found {len(windows)} opportunity windows for {sid}. "
            f"{len(high_score_wins)} window(s) have high suitability score (≥60). "
            "Ranked by block suitability score descending."
        )
        responses.append(OpportunityMiningResponse(
            section_id=sid,
            total_windows_found=len(windows),
            opportunities=windows,
            analysis_notes=notes
        ))

    return StandardResponse(data=responses, meta={"total_sections": len(responses)})