# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CorridorOps",
    "description": "CorridorOps - AI-Powered Automatic Block Planning System for Indian Railways (SIH PS #26027)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/corridor-events/synthesize": {
      "post": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Synthesize Events Endpoint",
        "operationId": "synthesize_events_endpoint_api_v1_corridor_events_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_CorridorEventSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/corridor-events": {
      "get": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "List Corridor Events",
        "operationId": "list_corridor_events_api_v1_corridor_events_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "section_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_CorridorEventSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/corridor-events/{id}": {
      "get": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Get Corridor Event Detail",
        "operationId": "get_corridor_event_detail_api_v1_corridor_events__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_CorridorEventDetailSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scoring/criticality": {
      "post": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Score Criticality Endpoint",
        "operationId": "score_criticality_endpoint_api_v1_scoring_criticality_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriticalityScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_CorridorEventSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/scoring/criticality/{event_id}": {
      "get": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Get Criticality Explainability",
        "operationId": "get_criticality_explainability_api_v1_scoring_criticality__event_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_CriticalityExplainabilityResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/optimize": {
      "post": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Optimize Schedule Endpoint",
        "operationId": "optimize_schedule_endpoint_api_v1_schedule_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_ScheduleSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/schedule/{schedule_batch_id}/shadow-block": {
      "post": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Shadow Block Endpoint",
        "operationId": "shadow_block_endpoint_api_v1_schedule__schedule_batch_id__shadow_block_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_ScheduleSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulate/what-if": {
      "post": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "What If Simulation Endpoint",
        "operationId": "what_if_simulation_endpoint_api_v1_simulate_what_if_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_WhatIfSimulationResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/schedule/monthly": {
      "get": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Get Monthly Schedule",
        "operationId": "get_monthly_schedule_api_v1_schedule_monthly_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Division"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_ScheduleSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/weekly": {
      "get": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Get Weekly Schedule",
        "operationId": "get_weekly_schedule_api_v1_schedule_weekly_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "section_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Id"
            }
          },
          {
            "name": "week_start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week Start"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_ScheduleSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/weekly/reoptimize": {
      "post": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Reoptimize Weekly Schedule",
        "operationId": "reoptimize_weekly_schedule_api_v1_schedule_weekly_reoptimize_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trigger_event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trigger Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_ScheduleSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule/{schedule_batch_id}": {
      "get": {
        "tags": [
          "Stage 2 & 3 AI Decision Core & Scheduling"
        ],
        "summary": "Get Schedule Batch Detail",
        "operationId": "get_schedule_batch_detail_api_v1_schedule__schedule_batch_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_ScheduleSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CriticalityScoreRequest": {
        "properties": {
          "event_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Ids"
          }
        },
        "type": "object",
        "title": "CriticalityScoreRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleOptimizeRequest": {
        "properties": {
          "horizon_type": {
            "type": "string",
            "title": "Horizon Type",
            "default": "weekly"
          },
          "date_range": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range"
          },
          "section_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Ids"
          }
        },
        "type": "object",
        "title": "ScheduleOptimizeRequest"
      },
...
      "StandardResponse_CorridorEventDetailSchema_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CorridorEventDetailSchema"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[CorridorEventDetailSchema]"
      },
      "StandardResponse_CriticalityExplainabilityResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CriticalityExplainabilityResponse"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[CriticalityExplainabilityResponse]"
      },
...
      "StandardResponse_List_CorridorEventSchema__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CorridorEventSchema"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[List[CorridorEventSchema]]"
      },
...
      "StandardResponse_List_ScheduleSchema__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ScheduleSchema"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[List[ScheduleSchema]]"
      },
...
      "StandardResponse_WhatIfSimulationResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WhatIfSimulationResponse"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[WhatIfSimulationResponse]"
      },
...
      "SynthesizeRequest": {
        "properties": {
          "section_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Ids"
          }
        },
        "type": "object",
        "title": "SynthesizeRequest"
      },
...
      "WhatIfSimulationRequest": {
        "properties": {
          "schedule_batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Batch Id",
            "default": "default"
          },
          "variation_pct": {
            "type": "number",
            "title": "Variation Pct",
            "default": 15.0
          },
          "iterations": {
            "type": "integer",
            "title": "Iterations",
            "default": 100
          }
        },
        "type": "object",
        "title": "WhatIfSimulationRequest"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}